+++ tests/test_pagination.py
@@ -6,7 +6,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.orm.session import Session
 from sqlalchemy.sql.expression import select
 
-from sqlalchemy_things.pagination import CountOffsetPage, CountOffsetPaginator
+from sqlalchemy_things.pagination import OffsetPage, OffsetPaginator
 
 
 @pytest.mark.asyncio
@@ -23,9 +23,9 @@ async def test_count_offset_page_async(
             for i in range(92)
         ])
         stmt = select(mapped_class)
-        paginator = CountOffsetPaginator()
+        paginator = OffsetPaginator()
 
-        page: CountOffsetPage = await paginator.get_async(
+        page: OffsetPage = await paginator.get_page_async(
             sqlite_async_session, stmt, 1)
         items = tuple(page.items)
         for item in items:
@@ -37,14 +37,14 @@ async def test_count_offset_page_async(
         assert len(items) == 10
         assert page.total_items == 92
 
-        page = await paginator.get_async(
+        page = await paginator.get_page_async(
             sqlite_async_session, stmt, 5)
         assert page.previous == 4
         assert page.next == 6
         assert len(tuple(page.items)) == 10
         assert page.total_items == 92
 
-        page = await paginator.get_async(
+        page = await paginator.get_page_async(
             sqlite_async_session, stmt, 10)
         assert page.previous == 9
         assert page.next is None
@@ -65,9 +65,9 @@ def test_count_offset_page_sync(
             for i in range(92)
         ])
         stmt = select(mapped_class)
-        paginator = CountOffsetPaginator()
+        paginator = OffsetPaginator()
 
-        page: CountOffsetPage = paginator.get_sync(
+        page: OffsetPage = paginator.get_page_sync(
             sqlite_sync_session, stmt, 1)
         items = tuple(page.items)
         for item in items:
@@ -79,14 +79,14 @@ def test_count_offset_page_sync(
         assert len(items) == 10
         assert page.total_items == 92
 
-        page = paginator.get_sync(
+        page = paginator.get_page_sync(
             sqlite_sync_session, stmt, 5)
         assert page.previous == 4
         assert page.next == 6
         assert len(tuple(page.items)) == 10
         assert page.total_items == 92
 
-        page = paginator.get_sync(
+        page = paginator.get_page_sync(
             sqlite_sync_session, stmt, 10)
         assert page.previous == 9
         assert page.next is None